metadata-cleaner/requirements.txt
Flask
Pillow
piexif
pypdf
python-docx
exifread
